singluar product table redefined
