+++ server/database/index.js
@@ -4,9 +4,8 @@ const Product = require('./models/Product');
 const Review = require('./models/Review');
 const Profile = require('./models/Profile');
 const User = require('./models/User');
-const ProductCategory1 = require('./models/ProductCategory1');
-const ProductCategory2 = require('./models/ProductCategory2');
-const ProductCategory3 = require('./models/ProductCategory3');
+const ProductCategory = require('./models/ProductCategory');
+
 // export models here
 module.exports = {
   db,
@@ -14,7 +13,5 @@ module.exports = {
   Product,
   Review,
   User,
-  ProductCategory1,
-  ProductCategory2,
-  ProductCategory3,
+  ProductCategory,
 };
