+++ server/database/utils/seed/data/productcategory.js
@@ -1,0 +1,56 @@
+const productCategory1 = [
+  {
+    pcid: 1,
+    name: 'Computers',
+  },
+  {
+    pcid: 2,
+    name: 'Typerwriters',
+  },
+  {
+    pcid: 3,
+    name: 'Data Devices',
+  },
+  {
+    pcid: 111,
+    name: 'Motherboards',
+  },
+  {
+    pcid: 112,
+    name: 'CPU',
+  },
+  {
+    pcid: 221,
+    name: 'Screen',
+  },
+  {
+    pcid: 212,
+    name: 'No Screen',
+  },
+  {
+    pcid: 311,
+    name: 'Calculator',
+  },
+  {
+    pcid: 312,
+    name: 'Translator',
+  },
+  {
+    pcid: 11,
+    name: 'Parts',
+  },
+  {
+    pcid: 21,
+    name: 'Mecanical',
+  },
+  {
+    pcid: 22,
+    name: 'Electronic',
+  },
+  {
+    pcid: 31,
+    name: 'Hand Held',
+  },
+];
+
+module.exports = productCategory1;
