--- server/database/utils/seed/data/productcategory1.js
@@ -1,16 +1,0 @@
-const productCategory1 = [
-  {
-    pcid: 1,
-    name: 'Computers',
-  },
-  {
-    pcid: 2,
-    name: 'Typerwriters',
-  },
-  {
-    pcid: 3,
-    name: 'Data Devices',
-  },
-];
-
-module.exports = productCategory1;
