--- server/database/utils/seed/data/productcategory2.js
@@ -1,20 +1,0 @@
-const productCategory2 = [
-  {
-    pcid: 11,
-    name: 'Parts',
-  },
-  {
-    pcid: 21,
-    name: 'Mecanical',
-  },
-  {
-    pcid: 22,
-    name: 'Electronic',
-  },
-  {
-    pcid: 31,
-    name: 'Hand Held',
-  },
-];
-
-module.exports = productCategory2;
