--- server/database/utils/seed/data/productcategory3.js
@@ -1,28 +1,0 @@
-const productCategory3 = [
-  {
-    pcid: 111,
-    name: 'Motherboards',
-  },
-  {
-    pcid: 112,
-    name: 'CPU',
-  },
-  {
-    pcid: 221,
-    name: 'Screen',
-  },
-  {
-    pcid: 212,
-    name: 'No Screen',
-  },
-  {
-    pcid: 311,
-    name: 'Calculator',
-  },
-  {
-    pcid: 312,
-    name: 'Translator',
-  },
-];
-
-module.exports = productCategory3;
